requirements.txt
langchain>=0.1.0
langchain-community>=0.0.1
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
pypdf>=3.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0
flask>=2.3.0
gunicorn>=20.1.0
werkzeug>=2.3.0